Handle LF-only ansi art
Most ansi-art has CRLF, but in case we get input with only LF,
treat LF as "move to beginning of next line".  That's partially
redundant with CR, which is "move to beginning of current line",
but I haven't seen any ansi art which uses LF to only mean
"move to this position on next line", so the new interpretation
should be safe and work for both examples seen.